analyze/requirements.txt
libclang
networkx
dotenv
google.generativeai
